Amend

+++ src/templates/galleries.rs
@@ -3,8 +3,8 @@ use super::*;
 #[derive(Boilerplate)]
 pub(crate) struct GalleriesHtml {
   pub(crate) inscriptions: Vec<InscriptionId>,
-  pub(crate) prev: Option<usize>,
-  pub(crate) next: Option<usize>,
+  pub(crate) prev: Option<u32>,
+  pub(crate) next: Option<u32>,
 }
 
 impl PageContent for GalleriesHtml {
